+++ NEWS.md
@@ -10,6 +10,7 @@ Version 0.16.3 - dev
 - CHANGE: move `MText.set_font()` to `MTextEditor.change_font()`
 - CHANGE: move `MText.set_color()` to `MTextEditor.change_color()`
 - CHANGE: move `MText.append_stacked_text()` to `MTextEditor.stacked_text()`
+- BUGFIX: DXF export of GROUP checks for deleted entities
 
 Version 0.16.2 - 2021-04-21
 ---------------------------